podzial VideoPlayer na Webcam i FilePlayer, akceleracja GPU i synchronizacja etykiet
