+++ compose.yaml
@@ -58,6 +58,14 @@ services:
     command: [ "sh", "-lc", "uvicorn src.app.app:create_app --host ${HOST:-0.0.0.0} --port ${PORT:-8000} --factory" ]
     stdin_open: true
     tty: true
+    deploy:
+      resources:
+        reservations:
+          devices:
+            - driver: nvidia
+              count: all
+              capabilities: [gpu]
+
 
   frontend:
     build:
@@ -118,6 +126,14 @@ services:
         condition: service_healthy
     stdin_open: true
     tty: true
+    deploy:
+      resources:
+        reservations:
+          devices:
+            - driver: nvidia
+              count: all
+              capabilities: [gpu]
+
 
 volumes:
   db_data:
